Implement `sertype` and `serdata` extension points
To help with creating language bindings the Cyclone DDS C library exposes the
sertype and serdata structs which you can use to associate callbacks that
Cyclone can call into to handle things like receiving samples and
serializing/deserializing data directly into the memory layout expected by your
language.

You usually need to associate additional data with each sertype and serdata. To
do so you can create a wrapper struct with a C layout whose first field is the
sertype or the serdata. The callbacks from Cyclone provide a pointer to the C
sertype or the C serdata which you can then reinterpret to reconstruct the
wrapper type and access any metadata you want to store in that wrapper. The
wrapper for each type is typed in T to propagate the type to the callbacks as
needed.

It is also important to note that the serdata (representing the serialized data)
has three kinds:
  - sample: a full-fat sample containing a real message sent by an application.
  - key: just the key data associated with a sample (e.g. sent by dispose).
  - none: never exposed in the C API and treated as a hard error if it is seen.

sertype ops:

- sertype::free: Frees the sertype wrapper struct. Note, the lifecycle for the
  sertype is fully handed over to the C library once the sertype is passed
  through the topic creation API.
- sertype::zero_samples: Zeroes the memory associated with the samples.
- sertype::realloc_samples: Reallocate samples. TODO: not implemented yet.
- sertype::free_samples: Free previously allocated samples.
- sertype::equal: Check if two sertypes are equal. This is based on the type
  name and with type discovery should also be based on the type support
  identifier.
- sertype::hash: Computes a hash for the sertype (should probably adhere to
  sertype1 == sertype2 => hash(sertype1) == hash(sertype2)).
- sertype::type_id: Compute the type identifier for the sertype.
- sertype::type_map: Compute the type map for the sertype.
- sertype::type_info: Compute the type info for the sertype.
- sertype::derive_sertype: Derive another sertype from the sertype.
- sertype::get_serialized_size: Get the max serialized size associated with the
  sertype.
- sertype::serialize_into: Serialize a given sample based on the sertype into a
  destination buffer.

serdata ops:

- serdata::free: Frees the serdata wrapper struct.
- serdata::eqkey: Check if the keys associated with two serialized samples are
  the same.
- serdata::get_size: Get the size of the serialized sample.
- serdata::from_ser: Pull out the serialized sample or key from the fragchain.
- serdata::from_ser_iov: Pull out the serialized sample or key from the iovec.
- serdata::from_keyhash: Pull out the serialized key from the keyhash. The
  keyhash is either an MD5 hash (and so cannot be reconstructed into a key) or
  the big-endian CDR serialized form of the key. The big-endian serialization
  requires the type to not force an MD5 keyhash and requires the max possible
  serialized size of the type to be <= 16. In that case the keyhash bytes can
  attempt to be deserialized into the key type.
- serdata::from_sample: Construct a serdata from a provided sample. It can be
  either a full sample or a key, and depending on the need the correct data
  needs to be serialized.
- serdata::to_ser: TODO
- serdata::to_ser_ref: Prepare the serialized data (if it has not been already)
  and increment the serdata reference count.
- serdata::to_ser_unref: Decrement the serdata reference count.
- serdata::to_sample: Copy the data of the deserialized sample or key to the
  provided pointer.
- serdata::to_untyped: Create an untyped sertype based on the provided typed
  serdata.
- serdata::untyped_to_sample: TODO
- serdata::print: Write the provided serdata to a string (up to a length limit)
  for logging purposes.
- serdata::get_keyhash: Pull out the keyhash from the provided serdata.
- serdata::from_psmx: TODO

The majority of the testing for these functions occurs later when validating the
reader/writer/topic implementations but things like sample retrieval, the
conversion from the C kinds to the Rust wrappers, and the `ref`, `unref`,
create, and free lifecycle for the sertype/serdata are tested. To validate this,
the only callback function that needs to be non-null is the sertype free
callback and so no-op `sertype_free_mock` is passed through.

Commit-type: feature
Scope: cyclonedds