Add participant
Introduces a `Participant` type representing a DDS participant. Each participant
is tied to the lifetime of its associated domain via Rust's lifetime system.
This ensures that the participant does not outlive the domain it participates
in. Provides a `new` function for default construction and a `builder` function
for incremental configuration.

Commit-type: feature
Scope: cyclonedds