Add subscriber/publisher
In the DDS entity hierarchy, subscribers and publishers sit between the domain
participant and the data readers or writers. Unlike readers and writers, they
are not associated with a specific data type. A subscriber/publisher is an
agglomerate that readers or writers can be created under. Actions performed on
the subscriber or publisher can be propagated to all associated readers or
writers.

Note, generic typing only comes into play at the reader and writer level, where
entities communicate samples of a concrete type `T`.

Commit-type: feature
Subject: cyclonedds