Add waitset
A waitset allows a thread to wait on multiple entities simultaneously, such as
readers, writers, or conditions. When one or more attached entities are
triggered, the wait on the waitset returns.

Entities can be attached with optional blobs, which are returned when the
waitset triggers. As a result, the waitset is generic over the blob attachment
type `A`. To create a waitset when no blobs are going to be attached, it should
be typed using the unit type (e.g., `WaitSet::<()>`).

Commit-type: feature
Scope: cyclonedds