Transition from `Sample<T>` -> `SampleOrKey<T>`
This is based on the introduction of a `Keyed<T>` trait that allows a
key to be associated with a data type. One issue is that now all samples
must be keyed. The solution to this right now is to specify that the key
is the uninhabited type `!` (which is unstable and thus the mostly
equivalent `std::convert::Infallible` is used as this will eventually
decay into `!` as an alias an will be swapped to `!` then). When the
uninhabited type is specified as the key the untyped `Topic::new()`
constructor can be used (which should avoid keyed samples being
generated).

Commit-type: feature
Scope: cyclonedds