Add sample or key container
Topics can be keyed or unkeyed. In some situations, such as disposes, you may
only receive the key of a sample. The other language bindings typically handle
this by returning the sample type with only the key fields filled in, with the
other fields default-initialized, and the `valid_data` flag in the sample info
set to false.

In this binding, the distinction between sample and key is made explicit with
the `SampleOrKey` type. It can still decay into the full sample type, allowing
you to write code in the same style as other bindings, but it also enables
handling the key-only case explicitly, where only the key fields are present.

Each `SampleOrKey` instance also carries its associated sample info.

Commit-type: feature
Scope: cyclonedds