Add trait for computing max CDR bounds
This is used when reconstructing a key from a keyhash. If the maximum serialized
size of a key is unbounded or exceeds 16 bytes, the keyhash is the MD5 hash of
its serialized CDR representation. However, if the maximum serialized size is
<=16 bytes, the keyhash is simply the big-endian CDR-serialized form of the key
(which is then possible to deserialize).

Types can be either bounded to a fixed serialized size or unbounded (e.g.,
strings or sequences). The `CdrBounds` trait provides an associated function for
computing this size and specifying the alignment. All built-in std types that
make sense in the context of DDS now implement this trait, including their
fixed-size array forms and nested tuples.

Most languages offer a type that maps easily to the unbounded sequence (e.g., a
resizable vector) and the array (e.g., a fixed-size array). They also tend to
provide a string type that naturally maps to an unbounded string. However,
choosing a Rust type for a bounded string and a bounded sequence is less
obvious. Should it come from a popular existing crate that provides bounded data
types? Or should it just be a const-generic wrapper over a string or a vector
and be provided directly in this crate? This remains an open question.

The long-term intent is to pair the `CdrBounds` trait with a procedural macro so
that user-defined types with keys can derive their CDR bounds via compile-time
code reflection.

Commit-type: feature
Scope: cyclonedds