Add key hash
The key hash of a key is derived from its big-endian serialized form. If the key
type's maximum serialized size fits within 16 bytes, the key hash is simply the
serialized bytes themselves (allowing the original key to be recovered from the
hash). Otherwise, the key hash is computed as the MD5 hash of the serialized
data.

Commit-type: feature
Scope: cyclonedds